updated the stepper to community contributor
