+++ src/pages/NewUsers.jsx
@@ -8,5 +8,4 @@ function NewUsers() {
     </div>
   );
 }
-
 export default NewUsers;
